tradewinds-py/pyproject.toml
[project]
name = "tradewinds-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cartopy>=0.24.1",
    "contextily>=1.6.2",
    "geopy>=2.4.1",
    "matplotlib>=3.10.3",
    "psycopg2-binary>=2.9.10",
]
